Simplify delta generator
